+++ CHANGELOG.md
@@ -9,6 +9,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- First release of `exception-recommender`
+- First release of `exception-recommender`.
 
 [Unreleased]: https://github.com/giantswarm/exception-recommender/tree/main